Refactor timer module
